requirements.txt
gradio
openpifpaf
Pillow
gdown
numpy
scipy
cmake
openpifpaf
opencv-python-headless